# PERF: avoid creating numpy array in groupby.first|last